added mentor task

+++ Mentor/EmployeeManagementServices/src/main/java/com/emp/Repo/IEmployeeRepo.java
@@ -10,6 +10,6 @@ import com.emp.Entity.Employee;
 @Repository
 public interface IEmployeeRepo extends JpaRepository<Employee, Integer>{
 	
-	List<Employee> findByUserid(Integer userid);
+	List<Employee> findByUserid(Long userid);
 
 }
